+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Master
+
+- Added `-[NSObject observeNotification:selector:]`
+
 ## 3.18.10 (2019-09-25)
 
 - Added iPad 7 for `-[UIDevice modelName]`